+++ src/main/java/com/danilo/roombooking/dto/UserResponseDTO.java
@@ -3,12 +3,11 @@ package com.danilo.roombooking.dto;
 import com.danilo.roombooking.domain.User;
 import com.danilo.roombooking.domain.authority.Authority;
 
-import java.math.BigInteger;
 import java.sql.Timestamp;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-public record UserResponseDTO(BigInteger id, String username, String fullName, String email,
+public record UserResponseDTO(Long id, String username, String fullName, String email,
                               Timestamp createdAt, Timestamp updatedAt, Timestamp lastLogin,
                               Integer failedLoginAttempts, Boolean enabled, Set<String> authorities) {
     public UserResponseDTO(User user) {
